requirements.txt
pyodbc
groq
dotenv
streamlit
plotly
pyyaml
pandas
numpy
tabulate